+++ R/import_wrappers.R
@@ -42,7 +42,8 @@
 #'  in columns. Can be coded either with one row per individual and two columns
 #'  per loci or two rows per individual and two columns per loci using the
 #'  rows_per_individual argument. Genotypes can be pretty much anything,
-#'  although missing genotypes must be coded as -9. Must have a .str extension.}
+#'  although missing genotypes must be coded as -9. Must have a .str extension 
+#'  and be consistantly whitespace delimited.}
 #'
 #'  Sample and snp metadata can also be provided via file path, and will be read
 #'  in using \code{\link[data.table]{fread}} \emph{with the default settings}
@@ -93,8 +94,8 @@
 #'  to snp genotypes when importing delimited snps.
 #'@param rows_per_individual numeric (1 or 2), default 2. Number of rows used
 #'  for each individual.
-#'@param marker_and_sample_names logical, default FALSE. If TRUE, assumes that a
-#'  header row of marker and sample/sample metadata names is present.
+#'@param marker_names logical, default FALSE. If TRUE, assumes that a
+#'  header row of marker is present.
 #'
 #'@aliases read_vcf read_ms read_delimited_snps read_genepop read_FSTAT convert_genlight convert_genind
 #'@name snpR_import_wrappers
@@ -158,13 +159,13 @@ read_plink <- function(file){
 #' @export
 #' @describeIn snpR_import_wrappers Import STRUCTURE data files.
 read_structure <- function(file, snp.meta = NULL, sample.meta = NULL, 
-                           rows_per_individual = 2, marker_and_sample_names = FALSE,
+                           rows_per_individual = 2, marker_names = FALSE,
                            header_cols = 0){
   if(!grepl("\\.str$", file)){
     stop("File extension is not .str. Please check that the correct file has been entered and rename if needed.\n")
   }
   return(import.snpR.data(genotypes = file, rows_per_individual = rows_per_individual, 
-                            marker_and_sample_names = marker_and_sample_names, 
+                            marker_names = marker_names, 
                             header_cols = header_cols, snp.meta = snp.meta, sample.meta = sample.meta))
 }